Use custom fields to look for label content
If custom field label doesn't start with #, look for label with for attribute and text content matching custom field label, and use for attribute to look for input field.

I tried to use custom fields for this web:
http://www.ruralvia.com/teruel/

When you click on "Acceso ruralvia" it displays 3 field login. I setup custom field with id for third input, but it works for few days and then stop working, because id and name are different.