add passkey implementation (frontend only)
use the existing wasm passkey implementation from 1password
(passkey-rs) compiled by proton to webassembly as the basis for adding
passkeys to the extension.  This commit adds all the mechanics of
passkeys but no backend storage.  Note that the overriding of
nagivator.credentials.get and .create only works on Firefox.  To get
it to work on Chrome, the manifest version will have to be updated to
v3 so the injector can use world: main to hijack the methods.  When
this happens, the injector will have to have a main and isolated
component that communicate using window events due to the new
restriction on main scripts not being able to use backend APIs.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>